--- src/tacky/expr_result.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-#[derive(Debug)]
-pub enum ExprResult {
-    PlainOperand(Expr),
-    DerefrencedPointer(Expr),
-}
