--- src/tacky/top_level.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-#[derive(Debug, PartialEq)]
-pub enum TopLevel {
-    Fun(Function),
-    Static(StaticVariable),
-}
